+++ test/ansi16.test.js
@@ -4,7 +4,8 @@ import { esc } from './utils/helpers.js';
 // import env variables to simulate ANSI 16 color space
 import './env/color-space.ansi16.js';
 
-import { hex, green } from '../src/index.mjs';
+//import { hex, green } from '../src/index.mjs'; // for debugging only
+import { hex, green } from 'ansis';
 
 describe('color space', () => {
   test(`convert truecolor to ANSI 16 color space`, () => {