+++ test/ansi256.test.js
@@ -4,7 +4,8 @@ import { esc } from './utils/helpers.js';
 // import env variables to simulate ANSI 256 color space
 import './env/color-space.ansi256.js';
 
-import { hex, ansi256 } from '../src/index.mjs';
+//import { hex, ansi256 } from '../src/index.mjs'; // for debugging only
+import { hex, ansi256 } from 'ansis';
 
 describe('color space', () => {
   test(`convert truecolor to ANSI 256 color space`, () => {