+++ test/env/color-space.ansi16.js
@@ -1,2 +1,2 @@
 // set env variable to simulate ANSI 16 color space
-process.env.TF_BUILD = 'true';
+process.env.COLORTERM = 'ansi';