+++ test/env/color-space.ansi256.js
@@ -1,2 +1,2 @@
 // set env variable to simulate ANSI 256 color space
-process.env.TEAMCITY_VERSION = 'true';
+process.env.COLORTERM = 'ansi256';