+++ test/no-color.test.js
@@ -3,7 +3,8 @@ import { expect, describe, test } from 'vitest';
 // import env variables to simulate no color
 import './env/color-space.mono.js';
 
-import ansis, { red } from '../src/index.mjs';
+//import ansis, { red } from '../src/index.mjs'; // for debugging only
+import ansis, { red } from 'ansis';
 
 describe('color space', () => {
   test(`ansis.isSupported()`, () => {